requirements.txt
tensorflow-gpu==1.14.0
torch==0.4.0
scipy==1.2.1
matplotlib==2.2.3
# ~umpy==1.16.4
torchvision==0.2.1
opencv_python==4.1.1.26
numpy==1.18.1
